Add MapsController tests
Unit and Integration tests were included in the test harness that cover
the Maps Controller functionality. The integration test checks for an OK
status returned at the /maps endpoint given valid query string paramater
values for lat and long. The unit test checks for the correct JSON
structure to be returned by the /maps endpoint.