Add Spring Data REST to Pricing Microservice
The dependcies and proper configuration was done in order to add Spring
data REST functionality to the Pricing Microservice.

The Price model was given the @Entity model for use by the Data Rest
dependency. This generates the CRUD API for the Price model and the
appropriate service communication layers.

The Price Repository was converted into an interface that implements the
Spring CRUDRepository interface. This creates the proper CRUD data
persistence and database communications layer.

An H2 in memory database is configured for development purposes.